typing-backend/requirements.txt
python-dotenv
fastapi
mysql-connector-python
uvicorn
pydantic
